Move junit-runner to validation pipeline
