--- modules/pipeline/junit-test-runner/src/test/java/org/cafejojo/schaapi/pipeline/testrunner/junit/test/EmptyTest.java
@@ -1,4 +1,0 @@
-package org.cafejojo.schaapi.pipeline.testrunner.junit.test;
-
-public class EmptyTest {
-}
