+++ modules/validation-pipeline/junit-test-runner/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    compile project(":validation-pipeline")
+
+    compile group: "junit", name: "junit", version: junit4Version
+}
