+++ modules/validation-pipeline/junit-test-runner/gradle.properties
