+++ modules/validation-pipeline/junit-test-runner/src/test/java/org/cafejojo/schaapi/validationpipeline/testrunner/junit/test/EmptyTest.java
@@ -1,0 +1,4 @@
+package org.cafejojo.schaapi.validationpipeline.testrunner.junit.test;
+
+public class EmptyTest {
+}
